+++ src/utils/pluralize/types.ts
@@ -1,0 +1,7 @@
+export type Dictionaries = {
+  irregular?: [string, string][];
+  plural?: [string, string][];
+  singular?: [string, string][];
+  uncountable?: string[];
+};
+export type PluralizeRule = 'singular' | 'plural';